Updated to "append" new scores to previous scores in local storage.
You were always previously overwriting 'scores' in local storage with just the most recent game.

So this code loads the previous games, then add the new scores to them in local storage.

There are other places in the app that may not work the way you expect now, but this at least saves multiple games to local storage.