Incorporated Tom's code and fixed some bugs. Did some CSS updates as well.
Ran into a bug where same players were being added into the player list as well as the results list. Used a for loop to iterate over previous players, if they exist, do not add, if they don't exist, add them to the list. I also used a .slice(-2) on the resultsPage to trim the game results to the last two results added to the results array.